refactor: 쿠폰 생성 및 메일 발송 패키지 분리(`coupon`)
