+++ src/main/java/com/coupop/fcfscoupon/coupon/model/Coupon.java
@@ -1,4 +1,4 @@
-package com.coupop.fcfscoupon.model;
+package com.coupop.fcfscoupon.coupon.model;
 
 public class Coupon {
