+++ src/test/java/com/coupop/fcfscoupon/CouponServiceTest.java
@@ -9,6 +9,7 @@ import com.coupop.fcfscoupon.fcfsissue.exception.CouponNotOpenedException;
 import com.coupop.fcfscoupon.fcfsissue.exception.CouponOutOfStockException;
 import com.coupop.fcfscoupon.fcfsissue.exception.EmailAlreadyUsedException;
 import com.coupop.fcfscoupon.fcfsissue.model.FcfsIssuePolicy;
+import com.coupop.fcfscoupon.fcfsissue.support.RequestTime;
 import com.coupop.fcfscoupon.testconfig.DatabaseSetUp;
 import com.coupop.fcfscoupon.testconfig.MailSenderConfig;
 import java.time.LocalTime;
