+++ src/test/java/com/coupop/fcfscoupon/testconfig/MailSenderConfig.java
@@ -1,6 +1,6 @@
 package com.coupop.fcfscoupon.testconfig;
 
-import com.coupop.fcfscoupon.model.CouponEmailSender;
+import com.coupop.fcfscoupon.coupon.model.CouponEmailSender;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
